python/requirements.txt
bosdyn_api-4.0.0-py3-none-any.whl
bosdyn_core-4.0.0-py3-none-any.whl
bosdyn_client-4.0.0-py3-none-any.whl
pygame
pyPS4Controller
spatialmath-python
onnxruntime
